icom CIV separated from external access

+++ rtlsdr.c
@@ -74,7 +74,7 @@ int init_rtlsdr()
         exit(0);
     }
     
-    unsigned long tqrg = (TUNED_FREQUENCY * (1000000L+RTL_TUNER_CORRECTION))/1000000L;
+    unsigned long tqrg = (TUNED_FREQUENCY * (1000000L+(float)RTL_TUNER_CORRECTION))/1000000L;
     retval = rtlsdr_set_center_freq(dev, (unsigned long)tqrg);
     if(retval != 0) printf("freqset= %d\n",retval);
     
@@ -135,7 +135,7 @@ void rtlsetTunedQrgOffset(unsigned int hz)
     printf("rtl hz : %d\n",hz);
     unsigned long qrg = TUNED_FREQUENCY - hz;
     printf("rtl qrg : %ld\n",qrg);
-    qrg = (qrg * (1000000L+RTL_TUNER_CORRECTION))/1000000L;
+    qrg = (qrg * (1000000L+(float)RTL_TUNER_CORRECTION))/1000000L;
     printf("rtl qrg corr: %ld\n",qrg);
     int retval = rtlsdr_set_center_freq(dev, (unsigned int)qrg);
     if(retval != 0) printf("freqset= %d\n",retval);
